Update : map page 구조 변경
